Implement first version of node injections
For both stressing of cpus, and execution of arbitrary commands.
Both work with all of ThorFi modes: Constant injection, bursts, and degradation.

There was a bit of trouble getting things to work with cgroup2. In the end
I skipped that, for greater backwards compatibility. It works with cgroup1,
cgroup2 is a different project.

The node injectors architecture mirrors the net injectors architecture for
consistency. I'm sure the ThorFi people thought about their structure,
even though I'm not a huge fan of it.

I think up next is adding the fault injector into the mainline API

Limitations:
- stress limited to seconds1
- I use stress-ng for load injections onto the CPU, so the minimum size of bursts is one second
- I'm... not convinced that adding load onto a CPU within a cgroup is a a sensible test. It's in now, and if someone finds something with it that's great, but for me it's quite clearly at the limit of what mininet can do as an emulator.