Add redirect fault
This includes the actual feature, a config, and associated tests,
for both a single and multiple protocols.
Syntax is similar to known: a->b for the target interface,
and b->c for the redirect-to interface (which needs to be on the same
host). The target interface is set in the type_args, and enriched like
the main config.

Commit also comes with some refactors, bugfixes, the usual.

Redirect is implemented using tc-mirred: A filter is added to the qdisc,
and the action mirred is added to that filter. I'm unaware of filters
that come wit probabilities, so all traffic is redirected, not just
a percentage. Implementation of percentages might be possible with eBPF, but
that's currently out of scope.

I also found out that tcpdump deletes filters (at least the mirred one), and that callind Mininet() has side effects. Finding that out was not fun.
ingress qdiscs also don't seem to respect their assigned handle, but fixing that was relatively straightforward. I'm debugging to files now,
I think that's a better setup than what I've done previously.