Implement logging
Logging runs in (yet another) async task. It's based on internal state,
 which each fault modifies as it injects/ejects its faults.
This comes with the limitation that logs are based on the assumed
state, not the actual one, but this approach also removes the need
to implement cli command output parsing, and a whole bunch of other
things, massively reducing the amount of effort that it took
to implement this.