Implement interface up/down command
This interfaces with however the linux host handles interfaces, so it's a bit
platform specific. ThorFIs solution was based on ifdown, but that doesn't work on ubuntu, so now it runs with ifconfig.

Of course, that limitation is not ideal - but comprehensive coverage for most networking stacks is not something I'll get into right now