Implement MostUsedLinkFaultController
Config syntax is the same as RandomLinkController, with the addition of a nodes_blacklist, which is also available for the RandomLinkFaultController.

MostUsedFaultController comes with AgnosticLink, a simple wrapper object we use to communicate information
about links outside of mininet.
The "most used" link is identified by reading the amount of
TX and RX packets that passed through an interface during the last run.

I believe this injector is a good example for a slightly more "intelligent" injector, and it shows which parts of implementation
are boilerplate, and which aren't.

Comes with an example testcase