Refactor all FaultControllers into a BaseController
BaseController takes care of a lot of overhead-ish tasks, like
communication between Controller and ControllerStarter,
passing of messages, etc. In theory this should make it
very simple to build new FaultControllers - but the tradeoff is that
the architecture as a whole becomes more complex.

The BaseController currently doesn't support the "mode" settings used by
RandomLinkFaultController and MostUsedLinkFaultController, because I'm not
sure whether that is universal enough. However, that could easily be added to the BaseController if needed.

Comes with a "logging only" fault controller, that also illustrates
what to implement.