Made docker work
This is really weird. I could have sworn that it didn't work when I last tested it, but I ran it again,
and the fix I had planned to implement yesterday but never
got to was in the code already.
It's a christmas miracle!

I believe there's use case for "I want to run x in the
container, but I don't know how docker exec works" that
could be built into our logging, but that's not a priority for now.
On to... multi-injects? Ufff.

Also fixed unrelated stuff, made node injects and MostUsedLinkFaultController docker compatible, improved docs