Add random mode for all protocol redirect fault
Turns out that one of the basic filters has a "random" variable we can use to choose random packets.

With that in theory this should have been very simple. But meta(random gt any number) always returns true, and the same construct with lt always returns false, unless a mask is given for the random.
Even though the documentation doesn't imply that as far as I can see?

And this isn't googleable at all. So going from "this should work" to "this does work" took longer than expected.
Actually, cut that. By now that's exactly what I expected.

Special thanks to Felix for diving into the tc source to find that out.